Extract SSE stream plumbing out of syncController
syncController mixed HTTP request handling with the SSE transport machinery —
setupSSE (headers, proxy-buffering hardening, 2KB padding, destroyed-socket
guard) and executeSyncTask (single-task guard, 5s keepalive, res.on("close")
client-disconnect cancellation, error→event mapping). That plumbing is a
distinct concern from parsing a request and delegating to a service.

Move both into controllers/sseStream.ts. The controller now imports
executeSyncTask and holds only handlers (273 -> 195 lines); the delicate
res-vs-req close semantics and destroyed-socket guards live in one focused
module. Behavior unchanged — the SSE-contract tests still exercise this path
through the route. 148 passing, lint clean, build OK.
